+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Create Braintree class to be PSR compliant
 * Coinbase is no longer a supported payment method. `PAYMENT_METHOD_NO_LONGER_SUPPORTED` will be returned for Coinbase operations.
 * Add `ApplePay` for web domain registration.
+* Add facilitated details to Transaction if present
 
 ## 3.23.1
 * Fix token generator return type - thanks @jjok!